move set_arguments_to_cmd definition into experiment.py
