+++ src/otx/cli/tools/cli.py
@@ -28,7 +28,6 @@ from .export import main as otx_export
 from .find import main as otx_find
 from .optimize import main as otx_optimize
 from .train import main as otx_train
-from .run import main as otx_run
 
 __all__ = [
     "otx_demo",
@@ -40,7 +39,6 @@ __all__ = [
     "otx_train",
     "otx_optimize",
     "otx_build",
-    "otx_run",
 ]
